improved the error messages for the numeric fields

+++ src/main/scala/forms/report/ReportFormModel.scala
@@ -33,7 +33,7 @@ case class PercentageSplit(
 
 case class PaymentHistory(
                            averageTimeToPay: Int,
-                           percentPaidWithinAgreedTerms: BigDecimal,
+                           percentPaidWithinAgreedTerms: Int,
                            percentageSplit: PercentageSplit
                          )
